Dark mode cloud login

+++ client/src/api/fetchImage.js
@@ -25,17 +25,42 @@ export default async (endpoint, bucket, imageUrl, forcedHeight) => {
 
     // Only check cache if we force the height.
     if (forcedHeight) {
-      // TODO: Make sure the item is actually a blob.
       blob = await localforage.getItem(imageUrl)
     }
 
-    if (blob === undefined || blob === null || blob === '') {
+    if (
+      blob === undefined ||
+      blob === null ||
+      blob === '' ||
+      !(blob instanceof Blob)
+    ) {
       blob = await new COS({ endpoint: endpoint }).getObject({
         Bucket: bucket,
         Key: imageUrl
       })
 
-      if (forcedHeight) {
+      if (!blob.type.startsWith('image/')) {
+        // If responce isn't an image, try 4 more times, 500, 1000, 2000, 3000.
+        await new Promise((resolve, _) => {
+          const recursiveRetry = (wait, tries) => {
+            setTimeout(async () => {
+              blob = await new COS({ endpoint: endpoint }).getObject({
+                Bucket: bucket,
+                Key: imageUrl
+              })
+              if (blob.type.startsWith('image/') || tries === 0) {
+                resolve()
+              } else {
+                recursiveRetry((4 - tries) * 1000, tries - 1)
+              }
+            }, wait)
+          }
+          recursiveRetry(500, 3)
+        })
+      }
+
+      // If still isn't an image don't save it.
+      if (forcedHeight && blob.type.startsWith('image/')) {
         blob = await shrinkBlob(blob, forcedHeight)
         // Let's only cache if the image is small.
         localforage.setItem(imageUrl, blob)
